# Conflicts:
#	internal/identity/store/user_registration_store.go
#	internal/identity/store/user_registration_store_test.go
#	internal/platform/store/migrations.go
#	internal/platform/store/migrations_test.go
#	web/default/assets_test.go
#	web/default/scripts/generate-public-pages.mjs
#	web/default/src/features/packages/components.tsx
#	web/default/src/features/packages/index.tsx
#	web/default/src/features/packages/package-plan-card.tsx
#	web/default/src/features/users/components/dialogs/user-blind-box-dialog.tsx
#	web/default/src/features/users/components/users-table.tsx
#	web/default/src/features/wallet/components/blind-box-card.tsx
#	web/default/src/features/wallet/components/blind-box-sidebar.tsx

+++ internal/platform/ratelimit/redis_limiter.go
@@ -66,7 +66,7 @@ func (rl *RedisLimiter) Allow(ctx context.Context, key string, opts ...Option) (
 
 type Config struct {
 	Capacity  int64
-	Rate      int64
+	Rate      float64
 	Requested int64
 }
 
@@ -76,7 +76,7 @@ func WithCapacity(c int64) Option {
 	return func(cfg *Config) { cfg.Capacity = c }
 }
 
-func WithRate(r int64) Option {
+func WithRate(r float64) Option {
 	return func(cfg *Config) { cfg.Rate = r }
 }
